+++ CHANGELOG.md
@@ -25,7 +25,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   [GEOD-1091](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1091)
 - Added filtering by time and season for detail page
   [GEOD-873](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-873)
-
+- Added the spinner to detail page while it loads
+  [GEOD-1109](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1109)
+  
 ### Changed
 - Organize the data sources and classify the pills by region on the left navigation of the explore page.
   [GEOD-1086](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1086)