+++ CHANGELOG.md
@@ -5,7 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
-## ## [3.0.0-alpha.3] - 2018-06-29
+## [Unreleased]
+
+### Added
+- Added Optional Info Buttons on the Explore Page
+  [GEOD-1090](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1090)
+
+## [3.0.0-alpha.3] - 2018-06-29
 **Important - Need to have database with parameters setup before release in all projects**
 
 ### Added